Move unfoldNew into factories
